+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 ### ✨ Features and improvements
 
 - *...Add new stuff here...*
+
 - Upgrade target from ES2017 to ES2019
 - Removed `_interpolationType` unused field (#264)
+- Added calculateCameraOptionsFromTo to camera [#1427](https://github.com/maplibre/maplibre-gl-js/pull/1427)
 
 ### 🐞 Bug fixes